Refactor classes to utilise more OOP
